Restructure the devguide for Jython
Two strategies have been applied, depending on how CPython specific was
the original file. Where content applies to Jython, with simple edits,
we keep the original file. Where extensive re-writing is needed we start
a new "companion" file (e.g. grammar_jy.rst) and keep the original for
reference.

The aim is to allow changes to be merged as the CPython devguide
develops.

index.rst is an exception as it is the conventional start page and
cannot be renamed even though change is extensive.